APICollector/APICollector/requirements.txt
asttokens==2.2.1
backcall==0.2.0
black==23.3.0
blurhash==1.1.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
CouchDB==1.2
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
idna==3.4
ijson==3.2.0.post0
importlib-metadata==6.2.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter-client==8.1.0
jupyter-core==5.3.0
Mastodon.py==1.8.0
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==23.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
python-magic==0.4.27
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
tweepy==4.13.0
typing-extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.15.0
